separation en fichier ca compile
